requirements.txt
# Root-level requirements — mirrors backend/requirements.txt
# (used by some PaaS platforms that look for requirements.txt in the project root)

uvicorn[standard]>=0.29.0
fastapi>=0.111.0
python-multipart>=0.0.9
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-ollama>=0.1.0
langchain-huggingface>=0.0.3
sentence-transformers==2.7.0
transformers==4.40.0
tf-keras==2.15.0
pypdf>=4.2.0
docx2txt>=0.8
faiss-cpu>=1.8.0
python-dotenv>=1.0.1
nest_asyncio>=1.6.0
pydantic>=2.7.0